various updates from updating canonical protobuf JSON mappings
The most salient of which is that flow.DocsAndBytes now uses u32 instead
of u64 for accumulations of count mass.

DocsAndBytes is a protobuf message only within the scope of a single
transaction. Thereafter it's serialized as JSON and reduced using `sum`
annotations, which work with u64. So the impact of this change is minor
so long as no single transaction has more than 4GB of document bytes
processed.

We may want to revisit this, perhaps by allowing coercion with
string-encoded numeric values within the `sum` annotation.

+++ crates/derive/src/pipeline/mod.rs
@@ -310,7 +310,7 @@ impl Pipeline {
         self.stats
             .transform_stats_mut(transform_index as usize)
             .input
-            .increment(body.len() as u64);
+            .increment(body.len() as u32);
 
         let uuid = uuid.unwrap_or_default();
         let flags = uuid.producer_and_flags & message_flags::MASK;
